Add CONST_PTR indirection
Previously, we'd put large constants directly in the expression nodes
array, which means that each node had to be as large as a
pointer (i.e. 64 bits wide). This is wasteful (we don't need that much
space for most nodes) but it also ensures that we can't compare
pointers accross runs, since address layout randomization will make
them different.

Furthermore, this simplifies tile development, since most x86-64
instructions can't accept 64-bit constants, so they'd have to be
loaded into a register using the 'movabs' instruction and reduce to
register-register instructions instead. Now that the 'large' and small
constants are separated, this edge case handling is no longer
neccessary and this results in simpler tile code.